Module bytecode_verifier::absint

Structs

BlockInvariant

Enums

BlockPostcondition

JoinResult

Traits

AbstractDomain
Trait for finite-height abstract domains. Infinite height domains would require a more complex trait with widening and a partial order.

AbstractInterpreter

TransferFunctions
Take a pre-state + instruction and mutate it to produce a post-state Auxiliary data can be stored in self.

Type Definitions

InvariantMap
A map from block id's to the pre/post of each block after a fixed point is reached.