Module bytecode_verifier::borrow_graph

This module defines the (acyclic) borrow graph for the type and memory safety analysis. A node in the borrow graph represents an abstract reference. Each edge in the borrow graph is labeled with a (possibly empty) sequence of label elements. A label element is either a field index or a local index or a struct index. An edge coming out of frame_root (see abstract_state.rs) is labeled with a sequence beginning with a local or struct index followed by zero or more field indices. An edge coming out of a node different from frame_root is labeled by a sequence of zero or more field indices. An edge in the borrow graph from a node other than frame_root represents a prefix relationship between the references represented by the source and sink of the edge. There are two kinds of edges---strong and weak. A strong edge from node a to b labeled by sequence p indicates that b is equal to the p-extension of a. Instead, if the edge was weak, it indicates that b is an extension of the p-extension of a.

Structs

BorrowGraph
A borrow graph is represented as a map from a source nonce to the set of all edges coming out of it.

Edge
A labeled edge in borrow graph

Type Definitions

Label
The label on an edge