Module bytecode_verifier::struct_defs

This module provides a checker for verifying that struct definitions in a module are not recursive. Since the module dependency graph is acylic by construction, applying this checker to each module in isolation guarantees that there is no structural recursion globally.

Structs

RecursiveStructDefChecker