Function bytecode_verifier::verifier::verify_script_dependencies

pub fn verify_script_dependencies<'a>(
    script: &VerifiedScript, 
    dependencies: impl IntoIterator<Item = &'a VerifiedModule>
) -> Vec<VMStatus>

Verifying the dependencies of a script follows the same recipe as VerifiedScript::new ---convert to a module and invoke verify_module_dependencies. Each dependency of 'script' is looked up in 'dependencies'. If not found, an error is included in the returned list of errors. If found, usage of types and functions of the dependency in 'script' is checked against the declarations in the found module and mismatch errors are returned.