Struct bytecode_verifier::verifier::VerifiedScript

pub struct VerifiedScript(_);

A script that has been verified for internal consistency.

This does not include cross-module checking -- that needs to be done separately.

Methods

impl VerifiedScript

pub fn new(script: CompiledScript) -> Result<Self, (CompiledScript, VMStatus)>

Verifies this CompiledScript, returning a VerifiedScript on success.

On failure, returns the original CompiledScript and a list of verification errors.

Verification of a script is done in two steps:

Convert the script into a module and run all the usual verification performed on a module

Check the signature of the main function of the script

This approach works because critical operations such as MoveFrom, MoveToSender, and BorrowGlobal that are not allowed in the script function take a StructDefinitionIndex as an argument. Since the module constructed from a script is guaranteed to have an empty vector of struct definitions, the bounds checker will catch any occurrences of these illegal operations.

pub fn into_module(self) -> (ScriptConversionInfo, VerifiedModule)

Returns the corresponding VerifiedModule for this script.

Every VerifiedScript is a VerifiedModule, but the inverse is not true, so there's no corresponding VerifiedModule::into_script function.

pub fn serialize(&self, buf: &mut Vec<u8>) -> Result<(), Error>

Serializes this script into the provided buffer.

This is merely a convenience wrapper around script.as_inner().serialize(buf).

VerifiedScript instances cannot be deserialized directly, since the input is potentially untrusted. Instead, one must go through CompiledScript.

pub fn as_inner(&self) -> &CompiledScript

Returns a reference to the CompiledScript within.

pub fn into_inner(self) -> CompiledScript

Returns the CompiledScript within. Conversion back to VerifiedScript will require going through the verifier again.